making more friendly for invocation as a class
when running this as a class instead of a standalone app,
self.get_browser() is invoked at class instantiation time
but during the execution of run(), self.browser.quit() is
invoked which removes the session.

If run() is being invoked from a loop, the next time around
it will throw exceptions and generally not be good

This commit moves the browser creation and teardown to occur
wholly within the run() method.

Also added better debufgging support for class-based invocation